Upgrade zgrab_schema lifetime_hint to Long()
IAW conversation w/ @zakird.

+++ zgrab_schema.py
@@ -203,7 +203,7 @@ zgrab_tls = SubRecord({
     "session_ticket":SubRecord({
         "value":Binary(),
         "length":Integer(),
-        "lifetime_hint":Integer()
+        "lifetime_hint":Long()
     })
 })
